+++ GraphsLogic/Link.cs
@@ -7,9 +7,9 @@ namespace GraphsLogic
     public class Link
     {
         public Node Node { get; set; }
-        public double Value { get; set; } 
+        public int Value { get; set; } 
 
-        public Link(Node node, double value = 1)
+        public Link(Node node, int value = 1)
         {
             Node = node;
             Value = value;
